Validate JSON before Creating PR
As documented in the create-pull-request repo, pull requests created automatically will not have actions dependent on the "pull_request" hook run. This means the validate script does not run for any pull request that is automatically made.

Just to be on the absolute safe side, after we make changes to the JSON we run it through the validate script before making the pull request.

We hope it will not be required, but it's better to be safe than sorry.

Read more: https://github.com/peter-evans/create-pull-request/issues/48